requirements.txt
python-dotenv==1.1.1
SQLAlchemy==2.0.44
psycopg[binary]==3.2.12
pytest==8.4.1
APScheduler==3.10.4
